+++ src/AlgorithmsInterface.jl
@@ -32,11 +32,13 @@ export step!, solve, solve!, solve_loop!
 # stopping criteria
 export StoppingCriterion, StoppingCriterionState
 export StopAfter, StopAfterIteration, StopWhenAll, StopWhenAny
+export DefaultStoppingCriterionState, StopAfterTimePeriodState, GroupStoppingCriterionState
 
-export is_finished, is_finished!, get_reason, indicates_convergence
+export is_finished, is_finished!, get_reason, is_active, indicates_convergence
+export get_active_stopping_criteria
 
 # Logging interface
-export LoggingAction, CallbackAction, IfAction, ActionGroup
+export LoggingAction, CallbackAction, IfAction, ActionGroup, StopReasonAction
 export with_algorithmlogger, emit_message
 
 end # module AlgorithmsInterface